@@ -28,7 +28,7 @@
 namespace filament {
 
 // update this when a new version of filament wouldn't work with older materials
-static constexpr size_t MATERIAL_VERSION = 63;
+static constexpr size_t MATERIAL_VERSION = 64;
 
 /**
  * Supported shading models
